Fix comment: revert creation of startVerification API
