+++ examples/integrated-flows-examples/src/verification/app.ts
@@ -55,7 +55,7 @@ dotenv.config();
 
   const startSmsVerificationFlow = async (phoneNumber: string) => {
     const requestData = Verification.startVerificationHelper.buildSmsRequest(phoneNumber);
-    const response = await sinch.verification.startVerifications.startSms(requestData);
+    const response = await sinch.verification.verifications.startSms(requestData);
     console.log('Verification request sent! Please check the SMS on your you phone to get the OTP.');
     const answers = await inquirer.prompt([
       {
@@ -74,7 +74,7 @@ dotenv.config();
 
   const startPhoneCallVerificationFlow = async (phoneNumber: string) => {
     const requestData = Verification.startVerificationHelper.buildPhoneCallRequest(phoneNumber);
-    const response = await sinch.verification.startVerifications.startPhoneCall(requestData);
+    const response = await sinch.verification.verifications.startPhoneCall(requestData);
     console.log('Verification request sent! Please answer to the phone call ans listen to the OTP.');
     const answers = await inquirer.prompt([
       {
@@ -91,7 +91,7 @@ dotenv.config();
 
   const startFlashCallVerificationFlow = async (phoneNumber: string) => {
     const requestData = Verification.startVerificationHelper.buildFlashCallRequest(phoneNumber);
-    const response = await sinch.verification.startVerifications.startFlashCall(requestData);
+    const response = await sinch.verification.verifications.startFlashCall(requestData);
     console.log('Verification request sent! Please check the phone number calling you.');
     const answers = await inquirer.prompt([
       {
@@ -112,7 +112,7 @@ dotenv.config();
     const requestData = Verification.startVerificationHelper.buildDataRequest(phoneNumber);
     let response;
     try {
-      response = await sinch.verification.startVerifications.startData(requestData);
+      response = await sinch.verification.verifications.startData(requestData);
     } catch (error: any) {
       console.log(`Impossible to process the seamless verification: ${error.data})`);
       return;
