+++ examples/simple-examples/README.md
@@ -151,16 +151,16 @@ yarn run numbers:regions:list
 
 | Service             | Sample application name and location                                                                                                                         | Required parameters                           |
 |---------------------|--------------------------------------------------------------------------------------------------------------------------------------------------------------|-----------------------------------------------|
-| Start Verifications | [./src/verification/start/start-sms.ts](./src/verification/start/start-sms.ts)                                                                               | `VERIFICATION_IDENTITY`                       |
-|                     | [./src/verification/start/start-phonecall.ts](./src/verification/start/start-phonecall.ts)                                                                   | `VERIFICATION_IDENTITY`                       |
-|                     | [./src/verification/start/start-flashcall.ts](./src/verification/start/start-flashcall.ts)                                                                   | `VERIFICATION_IDENTITY`                       |
-|                     | [./src/verification/start/start-data.ts](./src/verification/start/start-data.ts)                                                                             | `VERIFICATION_IDENTITY`                       |
+| Start Verifications | [./src/verification/verifications/sms/start-sms.ts](./src/verification/verifications/sms/start-sms.ts)                                                       | `VERIFICATION_IDENTITY`                       |
+|                     | [./src/verification/verifications/phonecall/start-phonecall.ts](./src/verification/verifications/phonecall/start-phonecall.ts)                               | `VERIFICATION_IDENTITY`                       |
+|                     | [./src/verification/verifications/flashcall/start-flashcall.ts](./src/verification/verifications/flashcall/start-flashcall.ts)                               | `VERIFICATION_IDENTITY`                       |
+|                     | [./src/verification/verifications/data/start-data.ts](./src/verification/verifications/data/start-data.ts)                                                   | `VERIFICATION_IDENTITY`                       |
 | Verifications       | [./src/verification/verifications/sms/report-with-id_sms.ts](./src/verification/verifications/sms/report-with-id_sms.ts)                                     | `VERIFICATION_ID` + `VERIFICATION_CODE`       |
 |                     | [./src/verification/verifications/sms/report-with-identity_sms.ts](./src/verification/verifications/sms/report-with-identity_sms.ts)                         | `VERIFICATION_IDENTITY` + `VERIFICATION_CODE` |
 |                     | [./src/verification/verifications/flashcall/report-with-id_flashcall.ts](./src/verification/verifications/flashcall/report-with-id_flashcall.ts)             | `VERIFICATION_ID` + `VERIFICATION_CLI`        |
 |                     | [./src/verification/verifications/flashcall/report-with-identity_flashcall.ts](./src/verification/verifications/flashcall/report-with-identity_flashcall.ts) | `VERIFICATION_IDENTITY` + `VERIFICATION_CLI`  |
-|                     | [./src/verification/verifications/callout/report-with-id_callout.ts](./src/verification/verifications/callout/report-with-id_callout.ts)                     | `VERIFICATION_ID` + `VERIFICATION_CODE`       |
-|                     | [./src/verification/verifications/callout/report-with-identity_callout.ts](./src/verification/verifications/callout/report-with-identity_callout.ts)         | `VERIFICATION_IDENTITY` + `VERIFICATION_CODE` |
+|                     | [./src/verification/verifications/callout/report-with-id_callout.ts](./src/verification/verifications/phonecall/report-with-id_callout.ts)                   | `VERIFICATION_ID` + `VERIFICATION_CODE`       |
+|                     | [./src/verification/verifications/callout/report-with-identity_callout.ts](./src/verification/verifications/phonecall/report-with-identity_callout.ts)       | `VERIFICATION_IDENTITY` + `VERIFICATION_CODE` |
 | Verification-status | [./src/verification/verification-status/verification-by-id.ts](./src/verification/verification-status/verification-by-id.ts)                                 | `VERIFICATION_ID`                             |
 |                     | [./src/verification/verification-status/verification-by-identity.ts](./src/verification/verification-status/verification-by-identity.ts)                     | `VERIFICATION_IDENTITY`                       |
 |                     | [./src/verification/verification-status/verification-by-reference.ts](./src/verification/verification-status/verification-by-reference.ts)                   | `VERIFICATION_REFERENCE`                      |
