+++ examples/simple-examples/package.json
@@ -153,14 +153,14 @@
     "sms:inbounds:list": "ts-node src/sms/inbounds/list.ts",
     "sms:inbounds:get": "ts-node src/sms/inbounds/get.ts",
     "verification:verifications:start-sms": "ts-node src/verification/verifications/sms/start-sms.ts",
-    "verification:verifications:start-callout": "ts-node src/verification/verifications/callout/start-callout.ts",
+    "verification:verifications:start-phonecall": "ts-node src/verification/verifications/phonecall/start-phonecall.ts",
     "verification:verifications:start-flashcall": "ts-node src/verification/verifications/flashcall/start-flashcall.ts",
-    "verification:verifications:start-seamless": "ts-node src/verification/verifications/seamless/start-seamless.ts",
+    "verification:verifications:start-data": "ts-node src/verification/verifications/data/start-data.ts",
     "verification:verifications:report-with-id:sms": "ts-node src/verification/verifications/sms/report-with-id_sms.ts",
-    "verification:verifications:report-with-id:callout": "ts-node src/verification/verifications/callout/report-with-id_callout.ts",
+    "verification:verifications:report-with-id:callout": "ts-node src/verification/verifications/phonecall/report-with-id_callout.ts",
     "verification:verifications:report-with-id:flashcall": "ts-node src/verification/verifications/flashcall/report-with-id_flashcall.ts",
     "verification:verifications:report-with-identity:sms": "ts-node src/verification/verifications/sms/report-with-identity_sms.ts",
-    "verification:verifications:report-with-identity:callout": "ts-node src/verification/verifications/callout/report-with-identity_callout.ts",
+    "verification:verifications:report-with-identity:callout": "ts-node src/verification/verifications/phonecall/report-with-identity_callout.ts",
     "verification:verifications:report-with-identity:flashcall": "ts-node src/verification/verifications/flashcall/report-with-identity_flashcall.ts",
     "verification:verification-status:verification-by-id": "ts-node src/verification/verification-status/verification-by-id.ts",
     "verification:verification-status:verification-by-identity": "ts-node src/verification/verification-status/verification-by-identity.ts",
