+++ examples/simple-examples/src/verification/verifications/data/start-seamless.ts
