+++ examples/simple-examples/src/verification/verifications/sms/start-sms.ts
@@ -6,7 +6,6 @@ import {
   printFullResponse,
 } from '../../../config';
 
-/** @deprecated see ../../start/start-sms.ts instead */
 (async () => {
   console.log('***************************');
   console.log('* StartVerification - SMS *');
