+++ packages/verification/src/rest/v1/index.ts
@@ -1,5 +1,4 @@
 export * from './callbacks';
-export * from './start-verifications';
 export * from './verification-status';
 export * from './verifications';
 export * from './enums';
