--- packages/verification/src/rest/v1/start-verifications/index.ts
@@ -1,2 +1,0 @@
-export * from './start-verifications-api';
-export * from './start-verifications-api.jest.fixture';
