+++ packages/verification/src/rest/v1/verification-service.ts
@@ -1,7 +1,6 @@
 import { SinchClientParameters } from '@sinch/sdk-client';
 import { VerificationStatusApi } from './verification-status';
 import { VerificationsApi } from './verifications';
-import { StartVerificationsApi } from './start-verifications';
 
 /**
  * The Verification Service exposes the following APIs:
@@ -10,8 +9,6 @@ import { StartVerificationsApi } from './start-verifications';
  */
 export class VerificationService {
   public readonly verificationStatus: VerificationStatusApi;
-  public readonly startVerifications: StartVerificationsApi;
-  /** @deprecated */
   public readonly verifications: VerificationsApi;
 
   /**
@@ -25,7 +22,6 @@ export class VerificationService {
    */
   constructor(params: SinchClientParameters) {
     this.verificationStatus = new VerificationStatusApi(params);
-    this.startVerifications = new StartVerificationsApi(params);
     this.verifications = new VerificationsApi(params);
   }
 
@@ -35,7 +31,6 @@ export class VerificationService {
    */
   public setHostname(hostname: string) {
     this.verificationStatus.setHostname(hostname);
-    this.startVerifications.setHostname(hostname);
     this.verifications.setHostname(hostname);
   }
 }
